server_backend/api-gateway/requirements.txt
Flask
python-dotenv
requests
gunicorn
flask-cors